Move upload file
